tests/archived/requirements.txt
# Testing dependencies for NBA Betting Model
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xvfb==3.0.0
pytest-html==4.1.1
pytest-benchmark==4.0.0
pytest-timeout==2.2.0

# Coverage reporting
coverage==7.3.2
codecov==2.1.13

# Code quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Security scanning
bandit==1.7.5

# Performance profiling
memory-profiler==0.61.0
line-profiler==4.1.1

# Data testing
great-expectations==0.18.5
pandas-profiling==3.6.6

# Mock data generation
faker==20.1.0

# HTTP mocking for API tests
responses==0.24.1
requests-mock==1.11.0

# Async testing
pytest-asyncio==0.21.1

# Parallel testing
pytest-xdist==3.5.0

# Test fixtures and factories
factory-boy==3.3.0